requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
configparser==7.0.0
execnet==2.1.1
idna==3.7
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pytest==8.3.2
pytest-xdist==3.6.1
requests==2.32.3
urllib3==2.2.2
